Add main document, preface
This adds the main document structure for the text, along with the
initial draft of the unnumbered preface chapter. The font chosen for the
book is baskervillef, a freely-available font imitation of the
Baskerville serif font that is available as part of TeXLive.

If you have issues building the project, try using XeLaTeX and adding
the required font to your system.